# performance: Remove cloning of parser