## Summary

@xunilrj analysis showed that cloning the whole parser considerably reduces the parsing performance. 

The two places where we cloned the parser now have been in

`try_parse_ts` and inside `paren_or_arrow_expr`. 

* `try_parse_ts`: `rewind` used to panic but it seems that the bindings refactoring removed the code that caused the panic. It's now safe to use `rewind`.
* `paren_or_arrow_expr`: No idea what that code was doing. Removing it doesn't regress coverage, nor any test. I also compared the errors when parsing the libraries and there's no difference as well. Let's re-add it when we now what it is used for (with proper tests and such)

Some parse time goodies

* `vue.global.prod`: From 291ms to 12.4ms
* `react.production`: From 754us to 355us
* `dojo`: From 12ms to 2ms
* `d3.min` from 3s to 20ms (wtf?) 
* `pixi.min` from 1.9s to `21ms`
* `three.min` from 2.7s to 28ms

Part of #1837

## Test Plan

Compared coverage, ran unit tests