@IWANABETHATGUY 2021-12-01
how should is use the new efficient parser, i use rslint_lint parser in my codebase like this 
```toml
rslint_parser = { git = "https://github.com/rome/tools", branch = "remove-cloning"}
```
but i can't get clearly speed up from this new perf branch
anything wrong ?