@MichaReiser 2021-12-01
> how should is use the new efficient parser, i use rslint_lint parser in my codebase like this
> 
> ```toml
> rslint_parser = { git = "https://github.com/rome/tools", branch = "remove-cloning"}
> ```
> 


Looks about correct. Did cargo pull the new dependency? 

> but i can't get clearly speed up from this new perf branch anything wrong ?

Impossible to say without knowing more about how you use the parser, what you're testing and measuring. I don't expect this to change much for small files because it only improves the parse time but not the other phases as well.