@IWANABETHATGUY 2021-12-01
> To expand on this a bit. Our coverage tests (parsing about 17k very small JS files) doesn't benefit from these improvements at all because the cloning isn't very expensive in these cases (a few events, but not a 100 thousands or millions of events).

thanks a lot.